Reads entities from all the lists
It takes a long time to read the files, but the process of checking all
entities is quite quick. In a sample with a 20K line input:

$ wc input.txt
  20109  311596 2653543 input.txt

It took more than one second to load the entities from the files, and 47ms to
loop through all tokens looking for them.

$ go run netagger.go
Time to load entities:              1.336963s
Time to look through tokens:        47.437ms
Total execution time:               1.40539s